refactor: extract keyboard shortcut orchestration from MainLayout
Move all keyboard shortcut definitions and handlers into a dedicated
useCommands hook, reducing MainLayout's concerns to UI composition
and lifecycle management. Also extract ShortcutsModal into its own
component within the same file.

Closes #27
